Move aggregations and catch JS errors
