+++ CHANGELOG.md
@@ -4,6 +4,21 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/) and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
+## [Unreleased]
+
+### Added
+
+- Switch between daily, weekly and yearly views
+
+### Improved
+
+- JS error handling with React error boundary
+
+### Fixed
+
+- Loading indicator when the sizes view is loading
+- Catch errors when the sizes view throws an error
+
 ## [1.5.0] - 2020-02-16
 
 ### Added
