+++ package.json
@@ -40,7 +40,7 @@
     "is-url": "^1.2.4",
     "micro": "^9.3.4",
     "microrouter": "^3.1.3",
-    "mongoose": "^5.8.9",
+    "mongoose": "^5.9.1",
     "node-fetch": "^2.6.0",
     "node-schedule": "^1.3.2",
     "normalize-url": "^5.0.0",
@@ -49,8 +49,9 @@
     "prop-types": "^15.7.2",
     "react": "^16.12.0",
     "react-dom": "^16.12.0",
-    "react-redux": "^7.1.3",
-    "react-use": "^13.26.0",
+    "react-error-boundary": "^1.2.5",
+    "react-redux": "^7.2.0",
+    "react-use": "^13.26.1",
     "redux": "^4.0.5",
     "redux-devtools-extension": "^2.13.8",
     "redux-thunk": "^2.3.0",
@@ -69,7 +70,7 @@
     "eslint": "^6.8.0",
     "eslint-plugin-import": "^2.20.0",
     "eslint-plugin-react": "^7.17.0",
-    "eslint-plugin-react-hooks": "^2.3.0",
+    "eslint-plugin-react-hooks": "^2.4.0",
     "eslint-plugin-react-native": "^3.8.1",
     "mocked-env": "^1.3.2",
     "nyc": "^15.0.0",
@@ -77,6 +78,7 @@
   },
   "ava": {
     "verbose": true,
+    "timeout": "20s",
     "environmentVariables": {
       "ACKEE_TRACKER": "custom name"
     }
