+++ src/aggregations/aggregateDailyViews.js
