+++ src/aggregations/aggregateMonthlyViews.js
