+++ src/aggregations/aggregateYearlyViews.js
